+++ src/OMS.Challenge/OMS.Challenge.Core/Models/CreateActivityRequest.cs
@@ -6,14 +6,14 @@ namespace OMS.Challenge.Core.Models
     {
         public int EmployeeId { get; set; }
 
-        public int ActivityTypeId { get; set; }
+        public byte ActivityTypeId { get; set; }
 
         public DateTime ActivityDate { get; set; }
 
         /// <summary>
         /// The distance traveled in km
         /// </summary>
-        public decimal Distance { get; set; }
+        public double Distance { get; set; }
 
         public bool Validate()
         {
